Migrate encoding related tests to Jest
Along the way introduced `readName` and `writeName` converters to allow
for testing in isolation. This will also make it easier to allow for
introducing configurable name converters, as was requested.

Furthermore, leaving test cases for allowed characters alone.. I believe
this could be tested more effectively with a property based testing
approach.